Update .config/scripts/setUpCITools.sh Update .config/zsh/init.zsh

+++ dot_config/scripts/executable_setUpCITools.sh
@@ -1,25 +1,33 @@
 #!/bin/bash
-echo "📲 installing xcode-select tools"
-xcode-select --install
+#🍻 Update brew once per day
+IS_CI_SET_UP_RUN="$HOME/.isCISetUpRun"
+# Run only if last update was over 24 hours ago
+if [ ! -f "$IS_CI_SET_UP_RUN" ]; then
 
-echo "Applying Homebrew environment settings..."
-eval "$(/opt/homebrew/bin/brew shellenv)"
+  echo "📲 installing xcode-select tools"
+  xcode-select --install
 
-echo "⏳rebuilt bat cache"
-bat cache --build
+  echo "Applying Homebrew environment settings..."
+  eval "$(/opt/homebrew/bin/brew shellenv)"
 
-echo "📊 setup sketchybar"
-curl -L https://github.com/kvndrsslr/sketchybar-app-font/releases/download/v2.0.28/sketchybar-app-font.ttf -o $HOME/Library/Fonts/sketchybar-app-font.ttf
-(git clone https://github.com/FelixKratz/SbarLua.git /tmp/SbarLua && cd /tmp/SbarLua/ && make install && rm -rf /tmp/SbarLua/)
+  echo "⏳rebuilt bat cache"
+  bat cache --build
 
-sketchybar --load-font "Symbols Nerd Font"
-sketchybar --load-font "Symbols Nerd Font Mono"
-sketchybar --load-font "CommitMono"
+  echo "📊 setup sketchybar"
+  curl -L https://github.com/kvndrsslr/sketchybar-app-font/releases/download/v2.0.28/sketchybar-app-font.ttf -o $HOME/Library/Fonts/sketchybar-app-font.ttf
+  (git clone https://github.com/FelixKratz/SbarLua.git /tmp/SbarLua && cd /tmp/SbarLua/ && make install && rm -rf /tmp/SbarLua/)
 
-brew services restart sketchybar
+  sketchybar --load-font "Symbols Nerd Font"
+  sketchybar --load-font "Symbols Nerd Font Mono"
+  sketchybar --load-font "CommitMono"
 
-echo "🐁 mise installation of tools"
-mise install
+  brew services restart sketchybar
 
-echo "🔒make bitwarden as ssh service"
-export SSH_AUTH_SOCK=/Users/daniel/.bitwarden-ssh-agent.sock
+  echo "🐁 mise installation of tools"
+  mise install
+
+  echo "🔒make bitwarden as ssh service"
+  export SSH_AUTH_SOCK=/Users/daniel/.bitwarden-ssh-agent.sock
+
+  touch "$IS_CI_SET_UP_RUN"
+fi
